# src/main/resources/swagger-ui/swagger.yaml
openapi: 3.0.3

info:
  title: API MS Agendamento SUS
  description: API responsavel pelo ciclo de vida do agendamento de consultas e publicacao de eventos de comparecimento
  version: 1.0.0

tags:
  - name: Agendamentos
    description: Operacoes de criacao, confirmacao e encerramento de agendamentos
  - name: StatusConsulta
    description: Operacoes de manutencao dos status de consulta
  - name: StatusNotificacao
    description: Operacoes de manutencao dos status de notificacao

paths:

  /agendamentos:
    post:
      tags:
        - Agendamentos
      summary: Criar novo agendamento
      operationId: criarAgendamento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CriarAgendamentoRequest'
      responses:
        '201':
          description: Agendamento criado com sucesso

  /agendamentos/{idAgendamento}/confirmacao:
    patch:
      tags:
        - Agendamentos
      summary: Registrar confirmacao do paciente
      operationId: registrarConfirmacaoAgendamento
      parameters:
        - name: idAgendamento
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 10
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmacaoAgendamentoRequest'
      responses:
        '200':
          description: Confirmacao registrada com sucesso

  /agendamentos/{idAgendamento}/status:
    patch:
      tags:
        - Agendamentos
      summary: Registrar status final da consulta
      operationId: registrarStatusConsulta
      parameters:
        - name: idAgendamento
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 10
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtualizarStatusConsultaAgendamentoRequest'
      responses:
        '200':
          description: Status da consulta atualizado com sucesso

  /status-consulta:
    post:
      tags:
        - StatusConsulta
      summary: Criar status de consulta
      operationId: criarStatusConsulta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusConsultaRequest'
      responses:
        '201':
          description: Status de consulta criado com sucesso

    get:
      tags:
        - StatusConsulta
      summary: Listar status de consulta
      operationId: listarStatusConsulta
      responses:
        '200':
          description: Lista de status de consulta
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StatusConsultaResponse'

  /status-consulta/{id}:
    put:
      tags:
        - StatusConsulta
      summary: Atualizar status de consulta
      operationId: atualizarStatusConsulta
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtualizarStatusConsultaRequest'
      responses:
        '200':
          description: Status de consulta atualizado com sucesso

    delete:
      tags:
        - StatusConsulta
      summary: Remover status de consulta
      operationId: removerStatusConsulta
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '204':
          description: Status de consulta removido

  /status-notificacao:
    post:
      tags:
        - StatusNotificacao
      summary: Criar status de notificacao
      operationId: criarStatusNotificacao
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusNotificacaoRequest'
      responses:
        '201':
          description: Status de notificacao criado com sucesso

    get:
      tags:
        - StatusNotificacao
      summary: Listar status de notificacao
      operationId: listarStatusNotificacao
      responses:
        '200':
          description: Lista de status de notificacao
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StatusNotificacaoResponse'

  /status-notificacao/{id}:
    put:
      tags:
        - StatusNotificacao
      summary: Atualizar status de notificacao
      operationId: atualizarStatusNotificacao
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtualizarStatusNotificacaoRequest'
      responses:
        '200':
          description: Status de notificacao atualizado com sucesso

    delete:
      tags:
        - StatusNotificacao
      summary: Remover status de notificacao
      operationId: removerStatusNotificacao
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '204':
          description: Status de notificacao removido

components:
  schemas:

    CriarAgendamentoRequest:
      type: object
      required:
        - cns
        - dataConsulta
        - especialidade
        - crmMedico
        - cnesLocal
      properties:
        cns:
          type: string
        dataConsulta:
          type: string
          format: date-time
        especialidade:
          type: string
        crmMedico:
          type: string
          example: CRM-SP-123456
        cnesLocal:
          type: string
          example: 1234567

    ConfirmacaoAgendamentoRequest:
      type: object
      required:
        - statusNotificacaoId
      properties:
        statusNotificacaoId:
          type: integer
          format: int64
          example: 2

    AtualizarStatusConsultaAgendamentoRequest:
      type: object
      required:
        - statusConsultaId
      properties:
        statusConsultaId:
          type: integer
          format: int64
          example: 3

    StatusConsultaRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: COMPARECEU

    AtualizarStatusConsultaRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: FALTOU

    StatusConsultaResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        status:
          type: string
          example: COMPARECEU

    StatusNotificacaoRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: CONFIRMOU_48H_ANTECEDENCIA

    AtualizarStatusNotificacaoRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: CONFIRMOU_24H_ANTECEDENCIA

    StatusNotificacaoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        status:
          type: string
          example: CONFIRMOU_48H_ANTECEDENCIA